restructured CLI
