+++ lib/bootstrap.rb
@@ -15,6 +15,7 @@ require 'miq/mixin_miq.rb'
 require 'mixin_git.rb'
 require 'mixin_api.rb'
 require 'mixin_settings.rb'
+require 'cli/cli.rb'
 require 'manageiq.rb'
 require 'domain.rb'
 require 'feature.rb'
